# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_igeargeek_sales-system-api_internal_models.ActiveArchivedStatus:
    enum:
    - active
    - archived
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusArchived
...
  github_com_igeargeek_sales-system-api_internal_models.Company:
    properties:
      address:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deleted_by:
        type: integer
      id:
        type: integer
      industry:
        type: string
      legal_name:
        type: string
      name:
        type: string
      notes:
        type: string
      revenue_size:
        type: string
      size:
        type: string
      status:
        $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.ActiveArchivedStatus'
      tags:
        items:
          type: string
        type: array
      tax_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
      website:
        type: string
    type: object
  github_com_igeargeek_sales-system-api_internal_models.CompanySizeOption:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deleted_by:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handlers.companyForm:
    properties:
      address:
        type: string
      industry:
        type: string
      legal_name:
        type: string
      name:
        type: string
      notes:
        type: string
      revenue_size:
        type: string
      size:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      tax_id:
        type: string
      website:
        type: string
    type: object
  internal_handlers.companySizeOptionForm:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    CRM/sales-pipeline API. This spec currently covers the
    Prospect stage, Pipeline stage, and Dashboard summary routes
    as a scaffold — the annotation pattern on those handlers
    (see internal/handlers/prospect_stages.go, pipeline_stages.go,
    dashboard.go) extends to the rest of internal/routes/routes.go
    as they're annotated. Run `go run github.com/swaggo/swag/cmd/swag@v1.16.6
    init -g cmd/api/main.go -o docs --parseDependency --parseInternal` from the
    repo root to regenerate docs/ after editing annotations (the two flags are
    required to resolve gorm.DeletedAt across package boundaries).
  title: Sales System API
  version: "1.0"
paths:
  /admin/company-sizes:
    get:
      description: Returns every configured Company size bucket (active + inactive),
        ordered by name.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.CompanySizeOption'
            type: array
      security:
      - BearerAuth: []
      summary: List company sizes
      tags:
      - admin/company-sizes
    post:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Company size fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.companySizeOptionForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.CompanySizeOption'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a company size
      tags:
      - admin/company-sizes
  /admin/company-sizes/{id}:
    delete:
      description: 'Admin-only. Soft-delete (is_active: false).'
      parameters:
      - description: Company size ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a company size
      tags:
      - admin/company-sizes
    patch:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Company size ID
        in: path
        name: id
        required: true
        type: integer
      - description: Company size fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.companySizeOptionForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.CompanySizeOption'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a company size
      tags:
      - admin/company-sizes
...
  /companies:
    get:
      description: 'Paginated, filterable Company list, each row annotated with last_activity_at
        (from any company-scoped Activity). Filters: status, tag, industry, search
        (name), stale_days, has_won_deal.'
      parameters:
      - description: active or archived
        in: query
        name: status
        type: string
      - description: Filter by Company tag
        in: query
        name: tag
        type: string
      - description: Filter by industry
        in: query
        name: industry
        type: string
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Filter to companies with no activity in N days
        in: query
        name: stale_days
        type: integer
      - description: Filter to companies with (or without) a won Deal
        in: query
        name: has_won_deal
        type: boolean
      - description: Sort field, prefix - for descending (created_at, name, industry)
        in: query
        name: sort
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated company list (data, page, per_page, total)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List companies
      tags:
      - companies
    post:
      consumes:
      - application/json
      description: Creates a Company. name is required; industry/size/revenue_size
        must each match an active configured option (see /admin/industries, /admin/company-sizes,
        /admin/revenue-sizes). domain is derived server-side from website.
      parameters:
      - description: Company fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.companyForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Company'
        "400":
          description: Invalid body or missing name
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a company
      tags:
      - companies
...
  /companies/{id}:
    delete:
      description: Soft-delete (AuditedModel) — recoverable via Restore/Trash below.
        Never a hard delete, since Deals/Contacts/Payments reference company_id.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Company not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a company
      tags:
      - companies
    get:
      description: Returns a single Company, including its last_activity_at.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Company'
        "404":
          description: Company not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a company
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Updates a Company. industry/size/revenue_size must each match an
        active configured option; domain is re-derived from website.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Company fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.companyForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Company'
        "400":
          description: Invalid body or invalid option
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Company not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a company
      tags:
      - companies
  /companies/{id}/restore:
    post:
      description: Un-deletes a soft-deleted Company.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Company'
        "403":
          description: Not Admin/Sales Manager
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deleted company not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Restore a deleted company (Admin/Sales Manager only)
      tags:
      - companies
...
  /companies/trash:
    get:
      description: Returns soft-deleted Companies.
      produces:
      - application/json
      responses:
        "200":
          description: Paginated company list (data, page, per_page, total)
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not Admin/Sales Manager
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List deleted companies (Admin/Sales Manager only)
      tags:
      - companies
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT, e.g. "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"